Initial Controller Attack and AttackPoller

+++ Consumer-Game/Assets/Scripts/HealthInterface.cs
@@ -6,7 +6,7 @@ public interface HealthInterface
 {
     // Start is called before the first frame update    void Initialize(); //Function without any arguments
     void InitializeHealth();
-    float health { get; set; } //A variable
-    float maxHealth { get; set; }
-    void ApplyDamage(float points); //Function with one argument
+    int health { get; set; } //A variable
+    int maxHealth { get; set; }
+    void ApplyDamage(int points); //Function with one argument
 }
