+++ Consumer-Game/Assets/Scripts/Player/PlayerController.cs
@@ -75,8 +75,8 @@ public class PlayerController: HealthInterface
     protected float initialGravityModifier = 1f;
 
     //Health
-    [HideInInspector] public float health { get; set; }
-    public float maxHealth  { get; set; }
+    [HideInInspector] public int health { get; set; }
+    public int maxHealth  { get; set; }
     public virtual void InitializeHealth()
     {
         
@@ -84,7 +84,7 @@ public class PlayerController: HealthInterface
         health = maxHealth;
     }
 
-    public virtual void ApplyDamage(float points)
+    public virtual void ApplyDamage(int points)
     {
         health = Mathf.Clamp(health - points, 0, maxHealth);
     }
